feat: modified unit tests
