+++ src/aboutPage/__tests__/AboutPage.test.tsx
@@ -4,6 +4,7 @@ import i18n from 'i18next';
 import { I18nextProvider } from 'react-i18next';
 
 import AboutPage from '../AboutPage';
+import TestLoginProvider from '../../common/test/TestLoginProvider';
 
 vi.mock('react-router-dom', async () => {
   const module = await vi.importActual('react-router-dom');
@@ -19,13 +20,19 @@ describe('About Page', () => {
   const renderLang = (language: string) => {
     i18n.changeLanguage(language);
     render(
-      <I18nextProvider i18n={i18n}>
-        <AboutPage />
-      </I18nextProvider>
+      <TestLoginProvider>
+        <I18nextProvider i18n={i18n}>
+          <AboutPage />
+        </I18nextProvider>{' '}
+      </TestLoginProvider>
     );
   };
   test('renders AboutPage without errors', () => {
-    const { container } = render(<AboutPage />);
+    const { container } = render(
+      <TestLoginProvider>
+        <AboutPage />
+      </TestLoginProvider>
+    );
     expect(container).toBeTruthy();
   });
   test('renders AboutPage component in Finnish', () => {
