+++ src/auth/components/oidcCallback/OidcCallback.tsx
@@ -1,10 +1,15 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { RouteChildrenProps } from 'react-router';
 import { useTranslation } from 'react-i18next';
+import {
+  LoadingSpinner,
+  LoginCallbackHandler,
+  OidcClientError,
+  User,
+} from 'hds-react';
 import * as Sentry from '@sentry/react';
-import { LoadingSpinner } from 'hds-react';
-/* 
-import authService from '../../authService';
+
+import useAuth from '../../useAuth';
 import { useErrorPageRedirect } from '../../../profile/hooks/useErrorPageRedirect';
 import styles from './OidcCallback.module.css';
 import { getLinkRedirectState } from '../../../profile/hooks/useHistoryListener';
@@ -14,50 +19,69 @@ function OidcCallback({
 }: RouteChildrenProps): React.ReactElement | null {
   const { t } = useTranslation();
   const redirectToErrorPage = useErrorPageRedirect();
+  const { endLogin } = useAuth();
+  const generigErrorString = 'authentication.genericError.message';
 
-  useEffect(() => {
-    authService
-      .endLogin()
+  const onSuccess = (user: User) => {
+    // Handle successful login - redirect to profile page
+    endLogin(user)
       .then(() => {
         history.replace('/', getLinkRedirectState());
       })
       .catch((error: Error) => {
-        // Handle error caused by device time being more than 5 minutes off
-        if (
-          error.message.includes('iat is in the future') ||
-          error.message.includes('exp is in the past')
-        ) {
-          redirectToErrorPage({
-            message: t('authentication.deviceTimeError.message'),
-          });
-        } else if (
-          // Handle error caused by end user choosing Deny in Tunnistamo's
-          // permission request
-          error.message ===
-          'The resource owner or authorization server denied the request'
-        ) {
-          redirectToErrorPage({
-            message: t('authentication.permissionRequestDenied.message'),
-          });
-        } else {
-          // Send other errors to Sentry for analysis
-          Sentry.captureException(error);
-          // Give user a generic error
-          redirectToErrorPage({
-            message: t('authentication.genericError.message'),
-          });
-        }
+        // Send other errors to Sentry for analysis
+        Sentry.captureException(error);
+        // Give user a generic error
+        redirectToErrorPage({
+          message: t(generigErrorString),
+        });
       });
-  }, [history, redirectToErrorPage, t]);
+  };
+
+  const onError = (error?: OidcClientError) => {
+    //Handle error
+    if (!error) {
+      redirectToErrorPage({
+        message: t('authentication.genericError.message'),
+      });
+      return;
+    }
+
+    // Handle error caused by device time being more than 5 minutes off
+    if (
+      error.message.includes('iat is in the future') ||
+      error.message.includes('exp is in the past')
+    ) {
+      redirectToErrorPage({
+        message: t('authentication.deviceTimeError.message'),
+      });
+    } else if (
+      // Handle error caused by end user choosing Deny in Tunnistamo's
+      // permission request
+      error.message ===
+      'The resource owner or authorization server denied the request'
+    ) {
+      redirectToErrorPage({
+        message: t('authentication.permissionRequestDenied.message'),
+      });
+    } else {
+      // Send other errors to Sentry for analysis
+      Sentry.captureException(error);
+      // Give user a generic error
+      redirectToErrorPage({
+        message: t(generigErrorString),
+      });
+    }
+  };
 
   return (
-    <div className={styles.wrapper}>
-      <LoadingSpinner small />
-      <p>{t('oidc.authenticating')}</p>
-    </div>
+    <LoginCallbackHandler onSuccess={onSuccess} onError={onError}>
+      <div className={styles.wrapper}>
+        <LoadingSpinner small />
+        <p>{t('oidc.authenticating')}</p>
+      </div>
+    </LoginCallbackHandler>
   );
 }
 
 export default OidcCallback;
-
-*/
