+++ src/userGuide/__tests__/UserGuide.test.tsx
@@ -4,6 +4,7 @@ import i18n from 'i18next';
 import { I18nextProvider } from 'react-i18next';
 
 import UserGuide from '../UserGuide';
+import TestLoginProvider from '../../common/test/TestLoginProvider';
 
 vi.mock('react-router-dom', async () => {
   const module = await vi.importActual('react-router-dom');
@@ -19,13 +20,19 @@ describe('User guide', () => {
   const renderLang = (language: string) => {
     i18n.changeLanguage(language);
     render(
-      <I18nextProvider i18n={i18n}>
-        <UserGuide />
-      </I18nextProvider>
+      <TestLoginProvider>
+        <I18nextProvider i18n={i18n}>
+          <UserGuide />
+        </I18nextProvider>
+      </TestLoginProvider>
     );
   };
   test('renders UserGuide without errors', () => {
-    const { container } = render(<UserGuide />);
+    const { container } = render(
+      <TestLoginProvider>
+        <UserGuide />
+      </TestLoginProvider>
+    );
     expect(container).toBeTruthy();
     expect(container).toMatchSnapshot();
   });
